Added controller and view for data reports
Used highcharts to display graphs. Graphs are currently hard coded, I'm
working to query the database now. Added tabs to the dashboard when not
signed in. Pages that require a registered user redirect to the sign in
page, but the data reports are accessible to anyone regardless of
whether or not they're signed in.